+++ src/components/settings/GeneralSettings.jsx
@@ -16,7 +16,7 @@ export function GeneralSettings({ accounts }) {
   return (
     <div>
       {/* Sub-tab navigation */}
-      <div className="flex border-b border-mail-border px-6 pt-2">
+      <div className="flex flex-wrap border-b border-mail-border px-6 pt-2">
         {generalSubTabs.map(sub => (
           <button
             key={sub.id}